Fix a require_once error during the extension installation
During the installation, there's a require_once error saying that the
"libphonenumber/PhoneNumberUtil.php" could not be open.

This file is inside the packages folder, which is added to the include_path by the extension's hook_civicrm_config, which is not called before the file is required.

To fix it, I moved that code out of the hook_civicrm_config and it will now be called whenever the extensions is available, including during the installation.